Added first set of integration tests
